Implemented player actions
Created enums Actions and Reactions to correspond to possible
player actions and reactions. Added an ArrayList of Actions as
a protected field to the Player class.

Created function setPossibleActions in the Player class to populate
the Player's actions ArrayList with the possible actions they can
perform, limited by their income.