Ambassador Exchange Implementation
Implemented Ambassador's exchange function, mainly through edits to a new class in Player called ambassadorExchange.

In Game, resolveTurnStack has a special exception for resolving Ambassador actions, because the Ambassador necessarily requires Player interaction with the Game unlike the other Actions.

ambassadorExchange is a large function mainly but can definitely be reduced in size. We should refactor the exchanges for one and two cards and also put the print display into a different function. Alternatively, we may make two different exchange functions which depend upon the number of hidden influences the player has.