Fixed reaction/action challenge bug
Fixed bug where a player who was challenged got to pick the card they wanted to reveal after revealing a card that failed a challenge. Correct behavior is to reveal the card that failed the challenge for them.

To do this, I overloaded the revealCard function to take in a CardType. In Game, we pass the CardType that failed the challenge to the revealCard function during the challenge resolution.